+++ crates/fuel-core/src/database/block.rs
@@ -32,7 +32,7 @@ use fuel_core_types::{
         },
         primitives::BlockId,
     },
-    entities::message::MerkleProof,
+    entities::relayer::message::MerkleProof,
     fuel_merkle::binary::MerkleTree,
     fuel_types::BlockHeight,
 };
@@ -166,7 +166,7 @@ mod tests {
                     },
                 };
                 let block = PartialFuelBlock::new(header, vec![]);
-                block.generate(&[])
+                block.generate(&[], Default::default())
             })
             .collect::<Vec<_>>();