+++ crates/types/src/entities/relayer.rs
@@ -1,0 +1,4 @@
+//! Relayed entities
+
+pub mod message;
+pub mod transaction;